pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "med-triage-env"
version = "0.1.0"
description = "MedTriage-Env: A medical diagnostic simulator for evaluating RL agents for OpenEnv Hackathon."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.1.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "usort>=1.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[tool.ruff]
line-length = 100

[tool.usort]
